requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
black==24.2.0
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.2
coverage==7.12.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
distinctipy==1.3.4
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
fairscale==0.4.13
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.61.0
fqdn==1.5.1
fsspec==2025.9.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.31
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.1
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
iopath==0.1.10
ipycanvas==0.14.2
ipykernel==7.1.0
ipympl==0.9.8
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
libcst==1.8.6
lightning-utilities==0.15.2
llvmlite==0.45.1
Markdown==3.10
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
moreorless==0.5.0
mpmath==1.3.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.5.0
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.1
pluggy==1.6.0
portalocker==3.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.10
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-rapidjson==1.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
ruff-api==0.1.0
safetensors==0.7.0
-e git+https://github.com/facebookresearch/sam3.git@757bbb0206a0b68bee81b17d7eb4877177025b2f#egg=sam3
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8
stack-data==0.6.3
stdlibs==2025.10.28
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.22
tinycss2==1.4.0
tokenize_rt==6.2.0
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchmetrics==1.8.2
torchvision==0.22.0+cu126
tornado==6.5.2
tqdm==4.67.1
trailrunner==1.4.0
traitlets==5.14.3
triton==3.3.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
ufmt==2.8.0
uri-template==1.3.0
urllib3==2.6.0
usort==1.0.2
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
widgetsnbextension==4.0.15
yacs==0.1.8
yt-dlp==2025.11.12
zstandard==0.25.0
